Move SnapshotUtility to appropriate folder
